Inital setup for Rabbit MQ integration.
