+++ docker-compose.yml
@@ -389,12 +389,22 @@ services:
     networks:
       - go-network
 
+  rabbitmq:
+    image: rabbitmq:3-management
+    container_name: rabbitmq
+    ports:
+      - 5672:5672
+      - 15672:15672
+    volumes:
+      - rabbitmq_data:/var/lib/rabbitmq
+
 volumes:
   user-management-db-data:
   stock-db-data:
   auth-db-data:
   stock-order-db-data:
   transaction-db-data:
+  rabbitmq_data:
 
 # ---------------------- Networks ----------------------
 networks:
